+++ src/brevitas_examples/llm/llm_quant/export.py
@@ -6,8 +6,12 @@ Copyright (C) 2023, Advanced Micro Devices, Inc. All rights reserved.
 from abc import ABC
 from abc import abstractmethod
 from contextlib import contextmanager
+from typing import Any
+from typing import Dict
+from typing import Iterable
 import warnings
 
+import gguf
 import numpy as np
 import torch
 
@@ -21,6 +25,26 @@ from brevitas.function.ops import max_int
 from brevitas.function.ops import min_int
 from brevitas.nn import QuantLinear
 from brevitas.proxy.parameter_quant import WeightQuantProxyFromInjector
+from brevitas_examples.llm.gguf_export.convert import ModelBase
+
+
+def find_hparam(keys: Iterable[str], hparams: Dict[str, int], optional: bool = False) -> Any:
+    key = next((k for k in keys if k in hparams), None)
+    if key is not None:
+        return hparams[key]
+    if optional:
+        return None
+    raise KeyError(f"could not find any of: {keys}")
+
+
+def gguf_tensor_map(self):
+
+    hf_arch = self.config.to_dict()["architectures"][0]
+    gguf_arch = ModelBase.from_model_architecture(hf_arch)
+    block_count = find_hparam(["n_layers", "num_hidden_layers", "n_layer", "num_layers"],
+                              self.config.to_dict())
+    tensor_map = gguf.get_tensor_name_map(gguf_arch.model_arch, block_count)
+    return tensor_map
 
 
 # TODO: Improve Groupwise export
@@ -39,6 +63,100 @@ def clip_kwargs(narrow, signed, bit_width):
             'max_val': max_int(signed, narrow, bit_width).to(dtype)}
     else:
         return None
+
+
+def gguf_mapping(state_dict):
+    updated_dict = dict()
+    tensor_map = gguf_tensor_map()
+    for k, v in state_dict.items():
+        if k.endswith('.attn_q_output'):
+            prefix = k.removesuffix('.attn_q_output') + '.q_proj'
+            suffix = '.q_output'
+        elif k.endswith('.attn_k_output'):
+            prefix = k.removesuffix('.attn_k_output') + '.k_proj'
+            suffix = '.q_output'
+        elif k.endswith('.attn_v_output'):
+            prefix = k.removesuffix('.attn_v_output') + '.v_proj'
+            suffix = '.q_output'
+        else:
+            prefix = k
+            suffix = ''
+        new_k = tensor_map.get_name(
+            prefix, try_suffixes=(".weight", ".bias", ".premul_input", ".q_input", ".q_output"))
+
+        assert new_k is not None, "Mapping to GGUF failed"
+
+        updated_dict[new_k + suffix] = v
+
+    return updated_dict
+
+
+def _optional_int_prop(p: dict[str, Any], name: str, default_value: int) -> int:
+    value = p.get(name, default_value)
+    try:
+        return int(value)
+    except ValueError as e:
+        raise ValueError(f"Property '{name}' expected to be an int and was not") from e
+
+
+def _float_prop(p: dict[str, Any], name: str) -> float:
+    try:
+        return float(p[name])
+    except ValueError as e:
+        raise ValueError(f"Property '{name}' expected to be a float and was not") from e
+    except KeyError:
+        raise KeyError(f"Property '{name}' not found (among keys {p.keys()})")
+
+
+def _get_dataset_props(config_json_struct) -> dict:
+    # Separate meta parameters (prefixed with _) from hparams.
+    meta_params = {k: v for k, v in config_json_struct.__dict__.items() if k.startswith("_")}
+    hparams = {k: v for k, v in config_json_struct.__dict__.items() if not k.startswith("_")}
+    return {
+        "meta": meta_params,
+        "hparams": hparams,}
+
+
+def _int_prop(p: dict[str, Any], name: str) -> int:
+    try:
+        return int(p[name])
+    except ValueError as e:
+        raise ValueError(f"Property '{name}' expected to be an int and was not") from e
+    except KeyError:
+        raise KeyError(f"Property '{name}' not found (among keys {p.keys()})")
+
+
+def convert_hf_hparams_to_gguf(hf_hparams: dict[str, any]) -> dict[str, any]:
+    hp = hf_hparams["hparams"]
+    attention_head_count = _int_prop(hp, "num_attention_heads")
+    attn_head_dim = int(_int_prop(hp, "hidden_size") // _int_prop(hp, "num_attention_heads"))
+    attn_head_dim = int(_optional_int_prop(hp, "head_dim", attn_head_dim))
+    return {
+        "llama.context_length":
+            _int_prop(hp, "max_position_embeddings"),
+        "llama.embedding_length":
+            _int_prop(hp, "hidden_size"),
+        "llama.block_count":
+            _int_prop(hp, "num_hidden_layers"),
+        "llama.feed_forward_length":
+            _int_prop(hp, "intermediate_size"),
+        "llama.rope.dimension_count":
+            attn_head_dim,
+        "llama.attention.head_count":
+            attention_head_count,
+        "llama.attention.layer_norm_rms_epsilon":
+            _float_prop(hp, "rms_norm_eps"),
+        "llama.attention.head_count_kv":
+            _optional_int_prop(hp, "num_key_value_heads", attention_head_count),}
+
+
+def _get_dataset_props(config_json_struct) -> dict:
+    # Separate meta parameters (prefixed with _) from hparams.
+    meta_params = {k: v for k, v in config_json_struct.items() if k.startswith("_")}
+    hparams = {k: v for k, v in config_json_struct.items() if not k.startswith("_")}
+    return {
+        "meta": meta_params,
+        "hparams": hparams,}
 
 
 class WeightBlockQuantHandlerBase(BaseHandler, ABC):
